+++ web/documentserver-example/csharp-mvc/Helpers/Users.cs
@@ -78,7 +78,8 @@ namespace OnlineEditorsExampleMVC.Helpers
             "Can't view chat",
             "Can't protect file",
             "View file without collaboration",
-            "Can’t submit forms"
+            "Can’t submit forms",
+            "Can’t refresh outdated file"
         };
 
         private static List<User> users = new List<User>() {