+++ web/documentserver-example/go/server/managers/default/user.go
@@ -129,6 +129,7 @@ var descriptionUser0 []string = []string{
 	"Can't protect file",
 	"View file without collaboration",
 	"Can't submit forms",
+	"Can't refresh outdated file",
 }
 var descriptionUser1 []string = []string{
 	"File author by default",