+++ web/documentserver-example/php/src/helpers/ExampleUsers.php
@@ -78,7 +78,8 @@ final class ExampleUsers
             "Can't rename files from the editor",
             "Can't view chat",
             "View file without collaboration",
-            "Can’t submit forms"
+            "Can’t submit forms",
+            "Can’t refresh outdated file"
         ];
         $this->users = [
             new Users(