+++ web/documentserver-example/ruby/app/models/users.rb
@@ -120,7 +120,8 @@ class Users
     "Can't view chat",
     "Can't protect file",
     'View file without collaboration',
-    'Can’t submit forms'
+    'Can’t submit forms',
+    'Can’t refresh outdated file'
   ]
 
   @users = [